# NibblePath GetHashCode performance